fix(equivalence): move baseline writer out of benchbox.cli (layering violation)
The new lint-imports check (import-linter, added by #952) enforces
utils < core < platforms < cli layering: core must never import cli. The
--update-baseline writer lived at benchbox/cli/cross_surface_baseline.py but
was imported from benchbox/core/equivalence/cross_surface.py, a real
violation caught fresh by CI on this PR.

The writer has no actual CLI concerns (no click decorators, invoked as a
plain function call) -- it only ever lived under cli/ by inherited
convention from the original AST-splicing implementation. Move it to
benchbox/core/equivalence/known_divergences_baseline.py, alongside the
cross_surface_baseline.yaml file it edits.
